+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-- N/A
+### Fixed
+
+- The `schema` prop will no longer be passed down to Ant Design and Chakra UI components.
 
 ## [v6.1.2] - 2023-03-21